# openapi/evry-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "currentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The user profile no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedUser"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The user profile no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/exercises/{id}": {
      "get": {
        "operationId": "ExercisesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDetail"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ExercisesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "Exercises"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/progress/overview": {
      "get": {
        "operationId": "ProgressController_overview",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "30d",
              "type": "string",
              "enum": [
                "30d",
                "90d",
                "6m",
                "1y",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressOverview"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/progress/activity": {
      "get": {
        "operationId": "ProgressController_activity",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "description": "Inicio inclusivo del rango; máximo 62 días hasta to.",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Final inclusivo; no puede superar el día actual en America/Bogota.",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressActivity"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/progress/exercises/{id}": {
      "get": {
        "operationId": "ProgressController_exerciseProgress[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "30d",
              "type": "string",
              "enum": [
                "30d",
                "90d",
                "6m",
                "1y",
                "all"
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaco; no combinar con page distinto de 1.",
            "schema": {
              "minLength": 1,
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 25,
              "default": 10,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseProgress"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ejercicio no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "progress"
        ]
      }
    },
    "/api/v1/progress/exercise/{id}": {
      "get": {
        "operationId": "ProgressController_exerciseProgress[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "30d",
              "type": "string",
              "enum": [
                "30d",
                "90d",
                "6m",
                "1y",
                "all"
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaco; no combinar con page distinto de 1.",
            "schema": {
              "minLength": 1,
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 25,
              "default": 10,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseProgress"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Ejercicio no encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "progress"
        ]
      }
    },
...
  },
  "info": {
    "title": "EVRY API",
    "description": "API privada de seguimiento de fuerza EVRY",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "evry_refresh"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer"
          },
          "avgPeriodLen": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "biologicalSex",
          "birthDate",
          "goals",
          "trackCycle",
          "avgCycleLen",
          "avgPeriodLen",
          "createdAt"
        ]
      },
      "UserUpdateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 date or date-time. Omitted or null values leave the current birth date unchanged."
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer",
            "minimum": 20,
            "maximum": 45
          },
          "avgPeriodLen": {
            "type": "integer",
            "minimum": 2,
            "maximum": 10
          }
        }
      },
      "UpdatedUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer"
          },
          "avgPeriodLen": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "biologicalSex",
          "birthDate",
          "goals",
          "trackCycle",
          "avgCycleLen",
          "avgPeriodLen"
        ]
      },
...
      "ExerciseDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sourceId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "muscleGroup": {
            "enum": [
              "CHEST",
              "BACK",
              "SHOULDERS",
              "BICEPS",
              "TRICEPS",
              "FOREARMS",
              "CORE",
              "QUADS",
              "HAMSTRINGS",
              "GLUTES",
              "CALVES",
              "FULL_BODY",
              "CARDIO"
            ],
            "type": "string"
          },
          "equipment": {
            "enum": [
              "BARBELL",
              "DUMBBELL",
              "MACHINE",
              "CABLE",
              "BODYWEIGHT",
              "KETTLEBELL",
              "BAND",
              "OTHER"
            ],
            "type": "string"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "bodyPart": {
            "type": "string",
            "nullable": true
          },
          "target": {
            "type": "string",
            "nullable": true
          },
          "secondaryMuscles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "equipmentLabel": {
            "type": "string",
            "nullable": true
          },
          "isCustom": {
            "type": "boolean"
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "isCompound": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "mediaId": {
            "type": "string",
            "nullable": true
          },
          "imagePath": {
            "type": "string",
            "nullable": true
          },
          "gifPath": {
            "type": "string",
            "nullable": true
          },
          "attribution": {
            "type": "string",
            "nullable": true
          },
          "instructions": {
            "nullable": true,
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ]
          },
          "instructionSteps": {
            "nullable": true,
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "gifUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sourceId",
          "name",
          "muscleGroup",
          "equipment",
          "category",
          "bodyPart",
          "target",
          "secondaryMuscles",
          "equipmentLabel",
          "isCustom",
          "ownerId",
          "isCompound",
          "tags",
          "description",
          "mediaId",
          "imagePath",
          "gifPath",
          "attribution",
          "instructions",
          "instructionSteps",
          "createdAt",
          "imageUrl",
          "gifUrl"
        ]
      },
...
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "ProgressPeriodWindow": {
        "type": "object",
        "properties": {
          "key": {
            "enum": [
              "30d",
              "90d",
              "6m",
              "1y",
              "all"
            ],
            "type": "string"
          },
          "from": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "timezone": {
            "type": "string",
            "enum": [
              "America/Bogota"
            ]
          }
        },
        "required": [
          "key",
          "from",
          "to",
          "timezone"
        ]
      },
      "OverviewMetrics": {
        "type": "object",
        "properties": {
          "sessionsCompleted": {
            "type": "integer",
            "minimum": 0
          },
          "volumeKg": {
            "type": "number",
            "minimum": 0
          },
          "activeDays": {
            "type": "integer",
            "minimum": 0
          },
          "weeklyFrequency": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "sessionsCompleted",
          "volumeKg",
          "activeDays",
          "weeklyFrequency"
        ]
      },
...
      "OverviewComparison": {
        "type": "object",
        "properties": {
          "previous": {
            "$ref": "#/components/schemas/OverviewMetrics"
          },
          "delta": {
            "$ref": "#/components/schemas/OverviewMetricDelta"
          }
        },
        "required": [
          "previous",
          "delta"
        ]
      },
      "ProgressRecord": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string"
          },
          "exerciseName": {
            "type": "string"
          },
          "kind": {
            "enum": [
              "WEIGHT",
              "REPS",
              "ESTIMATED_1RM"
            ],
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "achievedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "exerciseId",
          "exerciseName",
          "kind",
          "value",
          "achievedAt"
        ]
      },
...
      "ProgressOverview": {
        "type": "object",
        "properties": {
          "period": {
            "$ref": "#/components/schemas/ProgressPeriodWindow"
          },
          "summary": {
            "$ref": "#/components/schemas/OverviewMetrics"
          },
          "comparison": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OverviewComparison"
              }
            ]
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgressRecord"
            }
          },
          "muscleDistribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MuscleDistribution"
            }
          }
        },
        "required": [
          "period",
          "summary",
          "comparison",
          "records",
          "muscleDistribution"
        ]
      },
...
      "ProgressActivity": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgressActivityDay"
            }
          }
        },
        "required": [
          "from",
          "to",
          "days"
        ]
      },
...
      "ExerciseProgress": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string"
          },
          "period": {
            "$ref": "#/components/schemas/ProgressPeriodWindow"
          },
          "summary": {
            "$ref": "#/components/schemas/ExerciseProgressSummary"
          },
          "comparison": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ExerciseProgressComparison"
              }
            ]
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseProgressPoint"
            }
          },
          "history": {
            "$ref": "#/components/schemas/ExerciseProgressHistory"
          }
        },
        "required": [
          "exerciseId",
          "period",
          "summary",
          "comparison",
          "points",
          "history"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "retryable": {
            "type": "boolean"
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}